% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a quick box plot in ggplot.}
\usage{
gbox(x, y = "", color = "hotpink", fill = "lightpink", hz = FALSE)
}
\arguments{
\item{x}{This is vector to be plotted}

\item{y}{This is an optional categorical variable}

\item{color}{This is the color of the outline}

\item{fill}{This is the color that will fill in the boxplot (if categorical variable is given, the colors will also correspond to this option.)}

\item{horizontal}{TRUE or FALSE}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This will graph a given vector as a ggplot boxplot object
}
\examples{
## Create a box plot of one vector
library(datasets)
data(iris)
gbox(iris$Petal.Length)

## Create a Box plot of a numeric variable split by a categorical one.

gbox(iris$Petal.Width, iris$Species, hz = TRUE)


}
