% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{Create a histogram}
\usage{
ghist(x, color = "hotpink", fill = "lightpink", bins = 5)
}
\arguments{
\item{x}{This is the vector to be plotted.}

\item{color}{This is the color of the outline of the bars}

\item{fill}{This is the color of the filled in bars}

\item{bins}{This is the number of bins to make}
}
\value{
This function returns a ggplot histogram
}
\description{
This will graph a single numerical variable given as a vector.
}
\examples{
## Create a histogram of x
x <- rnorm(100, mean = 10, sd = 5)

ghist(x)

## Default of binwidth is 5. If we want more or less:
ghist(x, bins = 2)

}
